==> 1_pre_processing.ipynb <==
# Data Pre-processing

From the raw file that we got from rtf_conversion.ipynb, let's now pre-process the data into other csv's that we will be using for later methods, namely
1. colocates
2. topic modelling

For colocates, we want to have words still next to each other as would be found in the article text, but we do want to remove things like punctuation so we can segment each word and consider it at one "index". We should also remove general stopwords. 

For topic modelling, we are following a "bag-of-words" approach, so a lot more processing needs to be done. 

# general imports 
import pandas as pd
import re
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize, sent_tokenize
from nltk.stem import WordNetLemmatizer

# load dataframe of raw data
df = pd.read_csv('data/all_years_raw.csv')

# check if years have any issues
# there were no years detected above 2011, but there were for < 1980, 41 articles

df_test = df.loc[(df['year'] < 1980)] 
df_test

# a perusal of the data shows that the majority of these that were before 1980 are actually obituaries!
# so we will ignore them (remove them from the dataset)
df = df.loc[~(df['year'] < 1980)]

# check that they have been removed
df.loc[(df['year'] < 1980)]  

## Topic Modelling Section

The pre-processing for what we will need in the topic modelling

# Load stopwords
nltk.download("stopwords")
nltk.download("punkt")
stop_words = set(stopwords.words("english"))
nltk.download('wordnet')

# Define your cleaning functions
def remove_special_characters_and_numbers(text):
    return re.sub(r"[^a-zA-Z\s]", "", text)

def remove_stopwords_and_lowercase(text):
    word_tokens = word_tokenize(text)
    filtered_words = [w.lower() for w in word_tokens if w.lower() not in stop_words]
    return " ".join(filtered_words)

# Implement other cleaning functions (e.g., correct OCR errors, lemmatize/stem) if needed
def remove_urls_and_mentions(text):
    return re.sub(r"http\S+|www\S+|@\S+", "", text)

def remove_punctuation(text):
    return re.sub(r"[^\w\s]", "", text)

def lemmatize_text(text):
    lemmatizer = WordNetLemmatizer()
    return ' '.join([lemmatizer.lemmatize(word) for word in text.split()])

# additional stopwords
with open('data/stopwords-en.txt') as f:
    stopwords_txt = f.read()
    stopwords_list = stopwords_txt.split("\n")

# function to remove these additional stopwords
def remove_additional_stopwords(text):
    word_tokens = word_tokenize(text)
    filtered_words = [w for w in word_tokens if w not in stopwords_list]
    return " ".join(filtered_words)

# Apply cleaning functions to a new dataframe for topic modelling (df_tp)
df_tm = df 
df_tm = df_tm.replace(r'\n',' ', regex=True) 
df_tm['clean_text'] = df_tm['full text'].apply(remove_special_characters_and_numbers)
df_tm['clean_text'] = df_tm['clean_text'].apply(remove_stopwords_and_lowercase)
df_tm['clean_text'] = df_tm['clean_text'].apply(remove_additional_stopwords)
df_tm['clean_text'] = df_tm['clean_text'].apply(remove_urls_and_mentions)
df_tm['clean_text'] = df_tm['clean_text'].apply(remove_punctuation)
df_tm['clean_text'] = df_tm['clean_text'].apply(lemmatize_text)

print(df_tm.head())

# we get some sense of our data
df_tm['article_length'] = df_tm['clean_text'].apply(lambda x: len(x.split()))

# Basic statistical analysis
article_stats = df_tm['article_length'].describe()
print(article_stats)

# let us then export this df_tp as a .csv for use in the topic_modelling file
df_tm.to_csv('data_for_tm.csv', encoding='utf-8', index=False)

## Colocates section

Pre-processing that needs to be done for colocates

# we already ran the functions above in the topic modelling, so let's call the ones we need here
# we don't want to lemmatize the text though

df_cl = df 
df_cl = df_tm.replace(r'\n',' ', regex=True) 
df_cl['clean_text'] = df_cl['full text'].apply(remove_special_characters_and_numbers)
df_cl['clean_text'] = df_cl['clean_text'].apply(remove_stopwords_and_lowercase)
df_cl['clean_text'] = df_cl['clean_text'].apply(remove_additional_stopwords)
df_cl['clean_text'] = df_cl['clean_text'].apply(remove_urls_and_mentions)
df_cl['clean_text'] = df_cl['clean_text'].apply(remove_punctuation)

print(df_cl.head())

# we get some sense of our data
df_cl['article_length'] = df_cl['clean_text'].apply(lambda x: len(x.split()))

# Basic statistical analysis
article_stats = df_cl['article_length'].describe()
print(article_stats)

# let us then export this df_cl as a .csv for use in the colocate file

df_cl.to_csv('data_for_cl.csv', encoding='utf-8', index=False)
