==> 2_nltk_colocates.ipynb <==
# Colocates

It seems colocates can be spelled as co-locates, collocates, or colocates, so assume they are all the same. We used 'colocate' typically for the variable names and such.

This is done using the nltk library, with some math behind what happens


import pandas as pd
import nltk 
from nltk.collocations import *
from nltk.tokenize import word_tokenize

# import the pre-processed csv file 
df = pd.read_csv('data/data_for_cl.csv')

### Filtering for our analysis
Since we are doing analysis per decade, we have to filter out the years we want. This means that each time you run the file, you will have to double check which filter you want to apply 

we also filtered for two specific papers (Pittsburgh Post-Gazette (Pennsylvania), The New York Times) which is in the cell below the decade

# filter by decade 
# df = df.loc[df['year'] < 1990] # for 1980-1989
# df = df.loc[(df['year'] > 1989) & (df['year'] < 2000)] # for 1990-1999
# df = df.loc[(df['year'] > 1999) & (df['year'] < 2010)] # for 2000-2009

# filter by publisher (specifically NYT and pittsburg)
df = df.loc[df['publisher'] == 'The New York Times'] # 17833 articles 
# df = df.loc[df['publisher'] == 'Pittsburgh Post-Gazette (Pennsylvania)'] # 3229 articles

from gensim.utils import simple_preprocess

# change the cleaned text into a list
data = df['clean_text'].values.tolist()

# data_word lists 
data_words = [[word for word in simple_preprocess(str(doc))] for doc in data]

tokens = [i for j in data_words for i in j]
tokens

bigram_measures = nltk.collocations.BigramAssocMeasures()


word = 'feminism'

dict_col = {}
df_output_col = pd.DataFrame(columns=['word_of_interest', 'colocate', 'score'])

# Ngrams with 'feminism' as a member
word_filter = lambda *w: word not in w

## Bigrams - takes around 2-3 minutes per word
finder = BigramCollocationFinder.from_words(tokens)
# only bigrams that appear 3+ times
finder.apply_freq_filter(3)
# only bigrams that contain word of interest
finder.apply_ngram_filter(word_filter)
# return the 20 n-grams with the highest likelihood ratio
counter = 0 
for i in finder.score_ngrams(bigram_measures.likelihood_ratio):
    if counter < 20:
        print(i)
        df_output_col.loc[len(df_output_col.index)] = [word, i[0], i[1]] 
        counter += 1  
# print(finder.nbest(bigram_measures.likelihood_ratio, 15),)
# add it to a dictionary of lists 

df_output_col

# save the output as a csv for future use 
df_output_col.to_csv("col_nyt_feminism.csv")
